Forward port of fix:
Removing an incomplete hack and replacing it with a proper technique with
very little performance loss.

This hack tried to check that another repaint event was not issued around the
time that a scroll was triggered, as we would have to repaint the translated area
because the image on-screen had already been scrolled.  It was successful,
but only partially (the bug was only noticed less often in kdevelop).

This solution tracks which requests come from 3rd party applications, and
triggers an update of the whole view rather than trying to QWidget::scroll()
when that is the case.  It fixes it and is elegant :)

This also contains a slight change to the buffer highlighting tracking for
kio-accessed files, to make it more correct (there may have been a bug here
or not, we only got one unconfirmed report).

svn path=/trunk/kdelibs/kate/; revision=288068