One TODO less:
-  // TODO: genericise background painting

Painting for entire lines is aggregated in paintTextLineBackground since
aKademy, the painting of text blocks is now also generic.

It makes very little sense to combine these two as well, the version for
entire lines is a lot more efficient and does special processing that is
hard to do in the main loop (and not needed there).

At least there are now only two versions instead of the several slightly
different versions a week ago...

svn path=/trunk/kdelibs/kate/; revision=343820