XML autoindenter from Laurence Withers <lwithers@users.sourceforge.net> Laurence is willing to maintain it, so i suggest we provide him with a cvs account. Quoting his email with the patch: Working with Anders, I have developed an XML indentation mode. This mode will indent once per open element, so you end up with documents like:
<a>
  <b>
    <c>Some text and stuff <d>blah</d>
    </c>
  </b>
</a>

It ignores empty elements, processing instructions and CDATA sections
(these shouldn't affect the indentation). However, it also won't
recognise multi-line open or close tags (both of which are legal
according to the XML spec). This is a limitation caused by the use of
regular expressions.

I will maintain this mode as necessary, but it is useful as-is and I
would appreciate the feedback from users before making any further
changes.

svn path=/trunk/kdelibs/kate/; revision=350479