A large commit which is mostly concerned with adding an interface for model-based code completion.  It's functioning quite well at the moment, there are still several features to come (see kate/TODO), but at least the test seems to work :)
In regards to the other interface, the only major regression that I see
is the loss of ability to have multiple completion providers.  This is
mostly because the new way of starting a completion requires the client
code to provide a Range of text which the completion is applying to, and
I haven't figured out a good way to have multiple ranges.  Possibly this
can be worked around but I think it should actually be the client code's
responsibility to do any merging (as more information is available there
anyway).

Changes in detail:

KTextEditor:
* Document
  * replaceText function with non-string-list text parameter
* View
  * text cursor to widget coordinate function
  * added new position of the cursor to cursorPositionChanged
  * added mouse position functions and tracking enable/disable functions
* Code Completion
  * New model-based code completion interface.  Documentation coming
    shortly.

Kate:
* Part
  * clean up cursor position changed signal connections
  * commented out connection to argHintHidden() signal (doesn't exist?)
  * Renderer - changes to accomodate highlighting of out-of-document text
  * Highlighting - ditto
  * RenderRange -> KateRenderRange
  * View
    * Support new code completion interface
    * commented out connection of non-existant signals (old completion interface)
  * ViewInternal
    * support new arbitrary cursor coordinate function
    * support new code completion extension
  * Document - fixed modifiedOnDisk signal declaration
  * KateCompletion{Widget|Tree|Model|Delegate}
    * new classes implementing model-based code completion

* Test
  * Added code completion testing to katetest, using basic synthesised data

CCMAIL:kwrite-devel@kde.org
CCMAIL:ktexteditor-devel@kde.org

svn path=/trunk/KDE/kdelibs/interfaces/ktexteditor/; revision=502269