Prefix generated header files with ui_ for designer4 ui files. Adjust the includes accordingly (I hope I found all of them, at least kdelibs compiles here - except for kate/tests/arbitraryhighlighttest.cpp)
Be careful when using foo.ui, foo.h, foo.cpp now, though because uic puts
a very general include guard macro in the generated file: FOO_H
Most people (and kde-devel-vim.vim's ,i) will use the same macro for foo.h.
This will hopefully be fixable when Trolltech extends uic so that we don't
have to postprocess the output anymore. (The -o option of uic creates the
include guard from the name of the output file.)

svn path=/trunk/KDE/kdelibs/interfaces/ktexteditor/; revision=526900