Fix the problem with lines of text without a wrap point not being wrapped.
However, this is not the fix I thought would be needed.  I thought that
QTextOption::WrapAtWordBoundaryOrAnywhere would be the correct choice,
preferring to wrap at a word boundary if available, otherwise to wrap
anywhere (ie. when no more characters fit).  Instead, the option that
seems to work is QTextOption::WrapAnywhere, which does as the above
option describes.

It looks like QTextOption::WrapAtWordBoundaryOrAnywhere is only for
QTextDocument (looking at the source).  So, given it works, I'm happy.

svn path=/trunk/KDE/kdelibs/kate/; revision=533691