Clear exceptions in the script engine before calling indent, instead of after. Don't ignore the error, and try to convert an exception to an integer.
This has the same effect as the previous fix by zwabel, except that it doesn't potentially hide errors. Qt docs say that if calling indent throws an exception, the exception is returned. It doesn't make any sense to try and return this value as an indent amount. The real problem was that exceptions from previous calls were not cleared, which is fixed now.

svn path=/trunk/KDE/kdelibs/kate/; revision=814402