Rename the "win" parameter to "mainWin".
I don't remember the exact reason for this change
(it's like two months ago, or something?) but iirc
there's a potential situation when "win" is being used
for another variable in more complex plugins.

So this should be a good template and use the
non-ambiguous name. I think the original reasoning
was something like that.

svn path=/trunk/KDE/kdesdk/kate/; revision=833002