replace K3Process by KProcess
This currently does not even compile (it did not before) but this introduces no new compile errors. Collecting all stderr output of the process is rather straightforward.

If someone starts working on that and get's probles with the process stuff feel free to blame me for that (or ask me for fixing that).

svn path=/trunk/KDE/kdesdk/kate/; revision=864090