Code folding marker improvements: fix initial color, fix arrow luma, eliminate unneeded second color set.
This also removes the HSL blending code in favor of KCU::tint, which seems to work well. If not, we should add an HCY blend along the lines of ::tint (specifically, in how it handles the HC components for low-chroma colors, i.e. it should be ::tint but with linear Y response) to KCU.

TODO: make configurable!

BUG: 173695

svn path=/trunk/KDE/kdelibs/kate/; revision=908568