Allow the user to start more than one kate instance from the menu again.
First step to saner session handling.

*) The "new" command line option is ignored if the "start" option has an already opened session as parameter
*) Reuse of a specific instance is ignored, if the "start" option has an already opened session as parameter
==> the instance with the opened session is reused and raised

*) If no command line options are specified at all (except "block"), "new" is forced (you can start more than one kate from the menu again.
*) If no pid is set at all and the previous items did not catch, the "use" option is assumed, as was before

The other handling is done as as before.

I know it is a littlebit of a behavioral change, but in the end it is saner I guess.
The next planed step is, that the session chooser dialog switches to another already running instance,
if that instance has the requested session already open (except for default constructed sessions)

The end result should be:
easy access to more than one session
inhibiting "session loss" if more instances have the same session open

CCMAIL:kwrite-devel@kde.org

svn path=/trunk/KDE/kdesdk/kate/; revision=1023054