SVN commit 1203505 by coles:
Multi-line i18n calls in javascript using + for string concatenation breaks
gettext: only the first line of each of these messages was making it into
the .pot/.po files.  In lieu of an elegant fix (which may or may not exist)
the affected strings are now on one line.

CCMAIL: kde-i18n-doc@kde.org

SVN commit 1203502 by coles:

Minor proof-reading commit.  This string is not translatable yet beyond the
first line, so the change will not affect anything in the .pot/.po files.