If a dictionary has been set in the spell check dialog by the user, we use that one for the current language range in order to spell check *subsequent* words, which restores kdelibs < 4.4 spell checking behaviour. Currently, it is only used to influence the word suggestions that are presented to the user.
However, this shows a flaw in the spell check dialog as users seem to expect that
a dictionary change would apply to all the text in the document, whereas when the
dialog is shown some part of the text might already have been spell checked in
another language. To simplify matters, I think that the language combo box
should not be shown (i.e. it should be hidden for KatePart spell checking) in the
Sonnet dialog but this requires some changes to the Sonnet dialog in kdelibs.

BUG: 241647