renamed "Tree View" to "Documents" in settings & open settings from toolview popup.
*) It is now possible for plugins to register themselves with the toolviews they create.
If they do and implement the configuration page interface, the RMB menu of the buttons
(the one to switch the dock for the toolview) has a new action at the top "Configure ...".
This action triggers the first configuration page of the plugin

*) I didn't like "Tree View" in the configuration dialog. The toolview is called "Documents",
therefore I renamed the settings to "Documents" too. I think "Tree View" is too generic.