optimize: just draw a line for indentation-lines, instead of a many dots
Personally I never thought/saw that the indentation lines are supposed to be dotted.
To make things worse, the repetitive drawPoint calls where actually pretty slow,
slower even then text rendering in documents with many indentation levels

When we just draw a line, it's fast and looks still fine. the only difference
I see is that the dotted stuff had a bit less intensity due to its fake
alpha transparency by omitting every other pixel.

We could mimick that as well by desaturating the color or so if someone
really needs that, personally I'm fine with this "new" variant.