hidden option: disable line modification markers
In KDE >= 4.8.1, it is now possible to disable the line modification system
in Kate with

  kwriteconfig --file katerc --group "Kate View Defaults" --key "Line Modification" --type bool false

This only works if all Kate instances are closed. To enable it again, close
Kate and run

  kwriteconfig --file katerc --group "Kate View Defaults" --key "Line Modification" --type bool true

If you want to do the same for KWrite or Kile, use kwriterc or kilerc
instead of katerc. Same for KDevelop or other applications using kate part.

In master (KDE 4.9), there is a graphical option in "Appearance > Borders"

CCMAIL: kwrite-devel@kde.org
CCMAIL: kile-devel@lists.sourceforge.net
CCMAIL: lueber@gmail.com