extend document javascript api
new functions are:
- Range document.wordRangeAt(int line, int column)
- Range document.wordRangeAt(Cursor cursor)
Return the range of the word at the given cursor position. The returned range
is invalid (check range.isValid() in the script), if the text position is after
the end of a line. If there is no word at the given cursor, an empty range is
returned.

- bool document.wrapLine(int line, int column)
- bool document.wrapLine(Cursor cursor)
Wraps the line at the given cursor position.

Maybe backport to 4.8 branch for 4.8.3.

CCMAIL:kwrite-devel@kde.org