Add extra caret styles
"Underline" is a horizontal bar with a thickness of two pixels (the
same as for the default vertical bar). It is displayed when invoking
replace mode with "R".

"Half" is a "fat" underline, occupying half of the line height. It may
be used to indicate "operator-pending mode": reading a motion from the
keyboard after an operator command like "c", "d" or "y".