This is the first commit of the Pate plugin to the Kate repository.
All the sources are licensed for GPL V2 and V3, except the plugins
themselves (for which I am unclear what to do), and the .sip files
(which already have a copyright/license header).

The intent of this commit is to have a bare minimum of changes w.r.t.
the sources from github (i.e. the copyright/license headers, the
top level CMakeLists.txt, the README.txt and INSTALL.txt files).
Add Python scripting support using the Pate plugin to Kate's
playground.