Move caret style code out of KateRenderer::paintTextLine()
Each vi mode now sets the caret style itself. paintTextLine() does not
decide which caret to use, it only draws the given caret.

This commit also fixes a bug where the caret may not be drawn at all
when switching to vi input mode while the caret is flashing (the caret
flashes in regular mode, but not in vi mode). This is fixed by using
the repaint argument of KateView::setCaretStyle() to reset the timer.