Document the regular expressions
Composing regular expressions into larger expressions can result in
a large number of capture groups. Document the capture groups of the
larger expressions, and also add a PEG grammar outlining the command
syntax.

(There is a scaling problem here: it may be impractical to compose the
largest expressions into even larger expressions, since zero-or-one or
zero-or-more quantifiers can make the number of groups unpredictable.
One solution to this problem would be to define a function which
replaced all groups in a regular expression with shy groups.)