Disentangle configuration reading/saving (mostly).
There is a bit of a structural problem between the configuration of the
Pate plugin itself, of the Python modules which are wanted to be
loaded, and then their (Python) configuration. This is now largely
untangled.

A bit more clarification is probably warranted, especially with regard
to the "Apply" logic for the loading of modules.