Properly fix crashes due to moduleConfigPags being out of sync.
The previous commit was a temporary fix for a crash which was basically
due to the actual loaded modules being out of sync with the (cached)
list of module config pages after a reload.

We now reset the cache on a reload.

we also correct the reference counting for the stored tuple.