Make JavaScript commands work with vi ranges
If a command line script is called with a range -- e.g., "1,2sort" --
the command perceives the range as a selection, and only acts on those
lines. This is achieved by making the KateCommandLineScript class
inherit from RangeCommand in addition to Command. When
KateCommandLineScript acts as a RangeCommand, its exec() method sets
the selection to the given range.

(An alternative approach would be to handle vi ranges in the
JavaScript function each(), but this would not work for scripts
that access the selection directly. Abstracting the vi range as
a kind of selection makes for a simpler API.)