add DocumentCursor class
DocumentCursor extends Cursor by several functions to work with cursors
on a specific Document (e.g. wrapping cursor). However, it is not registered
like the MovingCursor, meaning that its position is not updated automatically.
DocumentCursor is much more efficient that MovingCursor, if you do not need
the automatic maintaining of the correct position in the document.

For KDE5, KTextEditor::DocumentCursor will be moved to KTextEditor :)

CCMAIL: kwrite-devel@kde.org