Restore exception propagaion for nested actions
With yesterday's changes, if action A was implemented in terms of
action B, and B raised an exception, it would get reported on exit from
B but would then not be propagated to A.

This change restores the control flow. It would be nice to only do the
reporting from A, represeting the full stack trace in the GUI, but
I have not worked out how to pull that off.