Allow "telescoping" mappings i.e. ones that are substrings of other mappings e.g.
'12
'123

etc

If the current set of keystrokes is a prefix of a mapping, then wait for the remaining keypresses that brings us up to that mapping. If that doesn't occur before the timeout occurs, choose the current mapping.  Ensure that we don't hang around waiting for timeouts if the current mapping is unambiguous.