kate pixmap scroll bar: take line decorations into account
this makes the feature work nicely with kdevelop.
it also shows the "highlight all" from searches in the bar, altough
the problem there is that as soon as you have more than one line per
pixel, some of the results might be skipped. This could eventually
be solved by adding a seperate overlay pixmap which is generated by
the search code. Scanning each line for highlights will probably be
too slow.