only remove trailing spaces after cursor
This commit is dedicated to Milian, since he seems to have a spasm
in his left hand, such that he presses CTRL+S all the time (this
disease seems to be very widespread). The story is like this:
If you remove trailing spaces on save (either of all or only of
modified lines), and the cursor position is in the trailing spaces,
then the cursor jumps to the last character. The workaround is
that only trailing spaces after the cursor position are removed.

This might leave trailing spaces in the saved document, so be
careful with your cursor navigation.

Be warned: All bug reports for this will be assigned to you.

CCMAIL: mail@milianw.de