minimap: Use foreground colour instead of a fixed one for handles
Previously, the scrollbar handles were drawn with (127,127,127,x), where x
was some alpha value. This worked better or less good, depending on your
colour scheme; for light schemes with white background color, it was
pretty intense, while for dark ones with background colors like (60, 60, 60)
it was barely visible. Now they're drawn with the foreground (text) color
(with a modified alpha value), which is guaranteed to give a good
contrast to the background.