Fix kate inputMethod query for Qt:ImCursorPosition and Qt:ImAnchorPosition
Current kate implementation for Qt:ImCursorPosition and Qt:ImAnchorPosition
is not correct. Qt:ImCursorPosition should always return the correct position
of cursor position of Qt::ImSurroundingText. Qt:ImAnchorPosition should return
the anchorPosition of selection, when in different line of cursor or there
is no selection, return the same position of cursor.

REVIEW: 104788