Include a synchronous wrapper for gdb.execute()
For scripting purposes, we need to be able to guarantee to get control
in the case of gdb.execute("run") etc. However, the obvious approaches
using threading don't work:

- gdb.execute("run") from a thread does not work. (GDB complains about
some difficulty fetching registers)

- Start Python timer, then gdb.execute("run") from the CLI also does
not work. (Process starts, and then hangs, but the timer never fires)

So, we resort to crude shell-based hackery to implement a timeout.