Revert "Programmer's move left/right" again
First of all, the feature does not work at all correctly at the moment.
Just try it, and it behaves rather non-deterministic, rendering it useless.

Next, the correct fix is to add an option in the "Editing > Navigation" config
tab, where the user can choose whether the cursor should stop at camelCaseNaming
and underscore.

Ideally, the implementation should be in our Cursor classes (WrappingCursor, ...)
and not copied twice without even really testing it. Sorry.

CCBUG: 46503