fix mouse-to-cursor icon-border-offset
Previously, coordinatesToCursor() always worked on view cordinates,
which is correct for KateView relative coordinates (i.e. with icon border),
and also for the CoordinatesToCursorInterface.

However, KateViewInternal::mouseMoveEvents() operates on coordinates
relative to the KateViewInternal, without the icon border. This regression
was incidently introduced by fixing bug 301990.

The bug is only present in betas/release canditates.

BUG: 312788
FIXED-IN: 4.10