Improve handling of URLs in vim mode
Patch by Fabian Kosmale:
Prior to this change, neither opening relative URLs like ../foo.txt nor
paths which included ~ were possible. This patch adds support for it. The
argument after :e will be interpreted as a relative URL if possible, else
as a absolute URL.
The relative URL is resolved relative to the URL of the active document if
it exists, else relative to the current working dir . This differs from
vim's  behaviour, where always the cwd is used. I decided for this,
because it was what was used  before, but have no strong opinion about
what should be used (active document URL, cwd or home dir as it used to
be).