build plugin: extend it so that it can handle many targets
Instead of a Build, Config, Clean and Quick command each target
now has a dynamic number of subtargets.
There is a new action select_target, which lists all available targets, and the user
can filter them.
The last built target is kept and can be built again via the new action build_previous_target.
Any target can be set to be the "Default" target, which is built via the normal run_make action,
and any target can be set to be the "Clean" target, which is built via make_clean.

The cmake clone at https://github.com/neundorf/CMake/tree/KateProjectGenerator can generate project
files for the project plugin which contains all the targets in the project etc.

Alex