Add test that selecting text "externally" to Vim (i.e. via the mouse, or one of the KTextEditor API's) switches us into Visual Mode with the given selection (already passes).
Add new KateViInputModeManager::isHandlingKeypress to check whether the current call to KateViVisualMode::updateSelection is caused by Vim mode or not; motivating testcase - where a mapping is triggered in Visual Mode, which indirectly causes us to switch from VisualLineMode to plain VisualMode in the call to KateViVisualMode::updateSelection - is included.

Also add another regression test for weird issue where a key sequence that mapped to "%", when used in Visual Mode, would end up causing us to yank too many characters when we pressed "y".  Looks like this has the same fix as the other bug.