Big chunk of work on incorporating ":" (i.e. running commands) in KateViEmulatedCommandBar. Includes basic running of commands (of course); display of output of commands; and completion of commands, along with the other advantages of KateViEmulatedCommandBar.  Also includes a bunch of refactorings.
In practice, I'm finding this a bit clunky, usability-wise: for example, one requires *two* presses of Enter to run the command: the first clears the completion box, the second actually runs the command.  This really needs to be changed to be more like the way Kate currently does things, which I'll do in due course.