Fairly big change that mainly changes the way the KateViEmulatedCommandBar deals with Enter when completion is active.  This enables us to do a :sort without having to press Enter twice. Completions are now filled in in the bar just by scrolling through the available completions, which is more in line with how Vim (and also Kate's own command bar) behave.  Cancelling a completion will revert to the last manual (i.e. - not arising from scrolling through the completion list) change.
Also fixes bug with Command completions not treating "-" as part of a command (e.g. in "set-auto-indent"), and also prevents completions popping up when we're writing e.g. :s/se<don't want completion here - it's annoying!>

As an experiment, I'm allowing completions invoked via ctrl-space to be dismissed via Enter without dismissing the whole bar; not sure if I'm happy with this, though, so I might change it.