We need to parse leading range expressions (e.g. "%", "'<,'>", ".,.+6" etc) in KateViEmulatedCommandBar when parsing a sed replace command (see motivating test cases), and the (fairly complex) code to do this is currently locked up in KateCmdLineEdit.
Rather than expose it via KateCmdLineEdit's (nicely minimal) API, it seems a better idea to just move it all into KateViEmulatedCommandBar: as Vim-specific code, it seems to belong there more.

Do this, and then use it with the ctrl-f & ctrl-d stuff to be able to clear the "find"/ "replace" portions of a sed expression even when the command has a leading range expression.

Vim-style ranges should still be usable in non-Vim mode.

REVIEW:111652