Hmmm ... Ok, not entirely happy with this, but here is a fix for some crashes when replaying a change/ executing a mapping.  Essentially, there is what is probably a bug in KateRenderer where, if a deletion is made and a paintEvent sent to KateViewInternal before its layout cache has been updated to reflect the document state after the deletion (this layout update is not done synchronously when a deletion is made), the part that draws trailing spaces can crash.
The use of m_viewInternal->repaint() in a couple of places triggered this very directly, but changing these to update() calls didn't help on its own, because we were calling sendPostedEvents inside feedKeyPresses, and this would again trigger paint events before the layout cache has been updated.  So use sendEvent instead.

For reference, the following non-Vim related code can trigger the crash, but I don't know if this is a valid use case so I'm hesitant to file a bug against KateRenderer:

    kate_document->setShowSpaces(true);
    kate_document->setText("ar ");
    kate_document->editBegin();
    kate_document->insertText(Cursor(0, 0), "b");
    kate_document->removeText(Range(0, 0, 0, 4));
    kate_view->parentWidget()->repaint();
    kate_document->editEnd();