Add an extension to Kate Vim in the form of a new shortcut ctrl-e for the emulated command bar, which behaves similarly to ctrl-r save that the inserted text is escaped for searching - this enables us to do e.g.
ygl

on the text:

( x / y * z * 5.0)

then open up a search and do ctrl-e, 0 and find all occurrences of that formula - in this particular case, the "/", "*" and "." are escaped when we do ctrl-e, 0 i.e. we search for

( x \/ y \* z * 5\.0)

This makes refactoring a duplicated complex expression like the above into a method/ local variable much more convenient.

DIGEST