A rather bigger commit than I would have liked :/
Anyway:

a) Tighten up textObjectAWord considerably, and make it behave more like Vim's, including all of Vim's rather weird corner bases that I can uncover.  In the process of doing this, I fell foul of some ambiguities in e.g. findPrevWordEnd etc, and so:

b) Have findPrevWordEnd etc return an invalid cursor if they can't find what we're asking them for.  This reduces some ambiguities, such as: if findPrevWordEnd returns Cursor(0, 0), does it mean that the previous word is a single character at the beginning of the document, or that there is no previous word at all? Updated all call sites, and added regression tests for each one.

c) Added textObjectAWORD and textObjectInnerWORD.  The latter is not as obsessively Vim-ish as the former.

d) Fix some debug info with the "run command with motion" stuff which printed out an incorrect range.

e) Made the special-casing of "w" and "W" inside the "run command with motion" part more specific: we're looking for *precisely* the "w" and "W" motions, not just a motion which ends in "w" or "W" (which includes the new "aw" and "aW").

There's some duplication between textObjectAWORD and textObjectAWord which I hope to remove.