Interactive sed replace when the 'c' flag is used - works only when a) in Vim mode and b) when also using the emulated command bar.
This removes sedMagic from SedReplace and replaces it with InteractiveSedReplacer, which is based on my recently re-worked sedMagic but with a few bug-fixes (e.g. it fixes a  bug where replacing text that originally contained newlines would not shift up the last line in the range, leading it to make replacement *beyond* the desired range) and also re-worked to be made, as the name suggests, interactive.  This could quite easily be used to implement interactive sed replaces in non-Vim mode.

DIGEST