This is a larger commit that I would have liked, but it's difficult to split up ("splitting up is hard to do", heh).
Essentially: I've finally figured out why reliably triggering shortcuts by sending the required keypresses didn't work: with some combinations of Qt and Xvfb, invoking/ dismissing the completion popup widget would deactivate our main window: since we were playing keypresses directly into it this did not have any real noticeable effect, except that it was no longer allowed to trigger shortcuts.  We work around this by introducing WindowKeepActive, which detects and reverses this deactivation.

Now that shortcuts can work, this exposed a few bugs that won't formerly being picked up by tests: mainly that dismissing a search bar while in Visual Mode could lead, for complex reasons, to a crash.  The fix for this was the sensible step of not letting the key mapper map the synthetic keypress sent by the dismissal of the search bar.  This in turn required (again, the sensible step) of better determining whether a keypress was synthetic or not.

I've now added an automated test for a bug that previously required a manual test, and have removed all the comments describing the manual test.  I've also added an automated test for detecting that the emulated command bar doesn't swallow all shortcuts intended for its QLineEdit which, again, was previously only testable manually.