A bit of a monster ;) Auto-completions cause trouble for Kate vim macros, especially in KDevelop; we may not get the same list of completions in the same order in the same circumstances due to e.g. C++ parsing errors, etc.  Additionally, completion of functions complicates matters: if there is an opening bracket after the cursor, then the completed function should *merge* its own opening bracket with the existing one; if there is not one, the completion should add one.  Therefore, we can't just blindly parrot the completion keypresses (ctrl-space; ctrl-p; ctrl-n; and Enter/ Return to select) we recorded when we replay the macro; the behaviour of  the replayed macro is context-sensitive.
Therefore, we now log completion events as they happen while recording a macro (and ditch the raw keypresses associated with choosing and triggering a completion), and make a note of some of the attributes (is it plain text? should it remove the tail? is it a function, and if so, should any opening brackets be merged?) and attempt to replay them in an intelligent manner.  I'll do the same for "." shortly.