searchbar: add message box when end of file is reached
This patch adds a message box (as in KDE 3) when the end/beginning
of the file has been reached while searching.
For those who don't like it, there is a "do not show again" checkbox.

For me, this improves usability a lot, since I otherwise never noticed
when the search had wrapped around, because I never noticed the
message on the status bar.

Alex