REVIEW: 113411 GUI:
Add an option to split STROUT and STDERR. If there smth in STDOUT, then
a passive error message will arrive w/ a command line end an error output.
When turned OFF, the textfilter plugin works as it was before:
output everything to a document despite of error.

Both options ("copy to clipboard" and "merge output") now are stored
in the configuration file.

If no output produced by the filter, the document remains unchaged.