swap file recovery: properly restore cursor position
Now undo and redo cursors are set correctly for the data that is
restored with the swap file recovery process. Thanks to this, you
can now undo/redo and see what text changed. This was not possible
before. So this is a nice improvement, since you now really know
what data is recovered, if you want this information.