KF5: Range normalization methods
This patch adds code, that is commented out, for normalizing a Range:
bool isNormalized(), void normalize(), Range normalized().

Usually, start() <= end() always holds. This is also guaranteed in the
constructor. However, currently, the API allows to set the end cursor by
calling range.end().setPosition(c), same for start(). So start() could
be > end() cursor.

In this case, normalizing the cursor makes sure that start <= end by
swapping end and start.

The other way is to change the API for KF5 such that we disallow setting
the start and end cursors by reference. Then, we'd have an API like
- Range::setStartPosition(c)
- Range::setEndPosition(c)
This way, the swapping would either be automatic in these functions, or
the cursor would simply not be set to c.

Any preferences?

CCMAIL: kwrite-devel@kde.org