Single characters made QChar instead of QString
Making single characters as QChar when they are added to a QString.
Adding single characters to a QString is faster if the characters are
QChars and not QStrings.

Same holds for arguments to QString::startsWith(), QString::endsWith(),
QString::remove() and QString::section()
Changes made to in these function calls as well

REVIEW: 113859