Force initialization of config variables.
This worked before but any checker would report access to
uninitialized memory when using KatePart then. The reason was that
when fooSet==true and we read the value from the config and it happens
to be the same as the randomly initialized value, no write takes place.
Thus valgrind is correct in assuming we access uninitialized memory
(we do). Thing is, we happen to have verified its validity.

To clean up this mess and make Katepart checkable with valgrind again,
we simply init all *Set variables to false in the global config ctors.

Note that no performance panalty can be observed, as at that point
no views or documents exist thus updateConfig is cheap.

OK'ed by Dominik.