Revert "Add Clojure support to symbolviewer."
This reverts commit fcbd9d0b6fe2e84afe0e903b6cf5e5e1d662980b.

There were several build breaks in this commit.
One of them was due to the use of Qt3Support (my Qt seems to be built
without it), which is a really really bad idea in new code.
Here is how to change the menu text properly:

m_func->setText()

CCMAIL: <r.c.g@gmx.de>