Fold KTextEditor and KatePart.
KTextEditor will stay the public API, KatePart the implementation.

This avoids to have always to jump through rings for any further extension and is the preparation
to have a sensefull KTextEditor Framework, that shall stay BC for KF 5.x and allows to use the
KTextEditor API without additional stuff required.