build plugin: add support for Intel icpc error messages
The format looks a bit different, an example is included in the
source file.
For gcc users still only one regex will be tested (after the first successfull one).
Intel users will have a bit more overhead.
I didn't measure it, but now the regex is checked only once, and after
that no additional searching in the string is done anymore, so it may
be even a bit faster.

Alex