Quick Open in tab bar: Show number of hidden documents
With this patch, the quick open button in the view space tab bar shows
the number of hidden documents by showing e.g. "+5" as button text. This
way, the users should always be aware that there are more documents open
that currently do not fit into the tab bar due to lack of space.

Further, this patch adds a What's This tip to the quick open button that
explains in detail what this is all about.