animate document tabs
This way, it is much easier to track tabs if a document is closed or
opened, or the main window is resized and therefore the tab bar's width
changes.

The size of the tabs follows browsers like FireFox or Chromium: If you
close a tab, the next tab moves under the mouse such that the close
button is exactly under the mouse again. Only after the mouse leaves the
tab bar, the tabs' width grows again. Feedback welcome.