build plugin: make "Build previous target" always build the previous target again
Before, only targets which had been selected explicitely, i.e. via
the selection-dialog or from the list, were considered.
I.e. if the default target or the clean target were built, these
were ignored when building the "previous target".
This was intended as a feature, so you could select the targets
you want and build them again using "build previous target",
and "make all" and "make clean" did not interfere with them.
But Kare considered it a bug, and actually I tend to agree,
so now also when building the default target, the next time
"build previous target again" is used, it will build the default
target (since this was the target which was built before).

Alex