fix/improve field highlighting in reStructuredText
Swap order of field and role rules, as the latter was overriding the
former and causing it to never fire. Change field detection to a regex
with a stricter rule as to what comprises a "field" (specifically, to
prevent a role at the beginning of a line from being interpreted as a
field). Add support for inline markup in fields.