rest.xml: don't consume extra inline markup char
Change 'inlineend' entity to a look-ahead assertion to avoid consuming
the character following inline markup. This fixes wrong highlighting of
such characters, as well as wrong parsing of e.g. "``literal``::".
Unfortunately there is not an easy solution for 'inlinestart', which has
the same problem (but would need a look-behind assertion, which are not
supported).