fix HL of indented reST code blocks
Add an additional context and dynamic rules in order to correctly detect
the end of a code block when the surrounding text is also indented. This
is especially important when parsing reST in Python docstrings, e.g.:

  """
  Example::
    This is the example.
  Before, this would still be in the code block context...
  """
  ...and this would still be reST (code, even) as a result.

(Note that reST HL support in python.xml is a local change on my machine
at the moment.)

The technique used here isn't my favorite, but there doesn't seem to be
a better way at present.